Add ProtocolCollection type to enforce whitelisted protocols
If a Protocol is in a ProtocolCollection we know it is supported.

Since the ProtocolCollection is reponsible for the whitelist, it probably makes sense for it to also be responsible for constructing protocols from a header version string.

Also remove explicit knowledge of allowed protocols from the parser, note:
- It is also no longer the responsibility of the parser to know how
to link the HEADER string to the protocol
- Because we now only accept protocol collections, the parser cannot be given nonsense in $allowedVersions – only valid protocol versions are accepted.