Improve clarity of documentation
There are no significant changes as far as active implementations are concerned.

The only thing "noteworthy" is the delineation between the version (which is format-agnostic) and the token (which is JSON-specific) in naming convention.

This is so that future designers might specify PASETO over an encoding format other than JSON (Protobuf, CBOR, etc.) without having to re-define the cryptography protocols. If this happens, `v4c` would be simply, "Version 4 with CBOR".

You can think of the tokens as `v4j` if that helps (although JSON gets "no suffix" reserved for it).

If you don't care about these kinds of future-proofing discussions, then this commit will be uninteresting to you.