Remove misleading implicit param from v1 and v2
This seems to have been added to aid unit testing with a generic
protocol.

Given the known expected behaviour for "implicit", allowing this
param to be specified (but ignored) seems like a fairly dangerous
footgun. For example, a user who has recently learnt about PASETO
in terms of v3 and v4 may specify this parameter during the
verification stage, expecting that it will be validated against the
string the specified during encryption. Unbeknownst to them, this
parameter will be ignored during both production and verification
of a PASETO.

An alternative to these changes is to throw an exception if
implicit is ever specified for v1 or v2.
I have chosen to gate this param though an interface extension
because it seems clearer to catch this through static analysis than
at runtime.