Error if received message is too short
Not making this check results in negative lengths, offsets, or both
being used in the destructuring of the payload. This can create
scenarios where certain payloads of different lengths are
indistinguishable, and where e.g. the exact same bytes are used in both
parts of the nonce, ciphertext, and mac.

This early abort does not interact with the plaintext and so no useful
information should be disclosed by this.